NewsPaper/requirements.txt
Django==3.1.7
celery==5.0.5
django-allauth==0.44.0
django-filter==2.4.0
django-modeltranslation==0.16.2

